Organize validation code into directory
