+++ pydian/validation/validate.py
@@ -3,7 +3,7 @@ from typing import Any
 from pydantic import create_model
 from result import Err, Ok
 
-from .dicts import get
+from ..dicts import get
 from .rules import Rule, RuleGroup
